[connman-jollagps] Don't use DBus for jolla-gps plugin. Contributes to JB#37470
This commit changes the jolla-gps plugin so that it no longer uses
DBus to communicate with the geoclue-hybris plugin process, but instead
writes the appropriate value to the location.conf file directly.
Note: this is not protected by a process-exclusion semaphore and so is
potentially dangerous!

It also removes the capability for the jolla-gps device power state
to be set by anything other than a transition to (or from) flight mode.